chore: the validation for email type already inserted
The validation have type email and is check if a email is valid.
So, minimum size was removed, if the person's email
is max@domain.com it would already be valid. Invalid emails would
not pass validation.
The maximum  size validation through a review was returned to
avoid sending emails with a quantity greater than that allowed in
the database.